reorganize helper modules for e2e tests
